Check if receiver hungup, assume handoff_concurrency
Now that handoff_concurrency enforces both outbound and inbound connections
there is a chance that the receiver will decide to reject the incoming handoff
and close the socket.  Since the current protocol has no way to express this
the sender must assume that a socket handup is the receiver rejecting the
handoff.

This is a normal occurance and should not be logged as an error/warning.